rename to dto and update content structure
