+++ apps/personal-site/src/app/content/pages/home.content.ts
@@ -1,4 +1,4 @@
-import { PageContent } from '../types/content.types';
+import { PageContent } from '../types/content.dto';
 
 export const HOME_CONTENT: PageContent = {
   title: 'Bryant Franks',
